+++ src/page-layout/AuthLayout/component/AuthForm/SignInForm.module.scss
@@ -1,0 +1,6 @@
+.formBox {
+  display: flex;
+  flex-direction: column;
+  gap: 2.8rem;
+  max-width: 35rem;
+}
