+++ src/page-layout/AuthLayout/component/index.ts
@@ -1,0 +1,1 @@
+export { default as SignInForm } from "./AuthForm/SignInForm";
